backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
coverage==7.11.0
cryptography==46.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.2.7
django-allauth==0.61.1
djangorestframework==3.16.0
idna==3.10
iniconfig==2.3.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0